% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aupdate.R
\name{Aupdate}
\alias{Aupdate}
\title{Aupdate}
\usage{
Aupdate(Dfix, Gamma, Adata, R, p, q, I, iniA, eps)
}
\arguments{
\item{Dfix}{the coefficient matrix for study covariates}

\item{Gamma}{the G matrix value}

\item{Adata}{the dataset}

\item{R}{the rank of reduced rank model}

\item{p}{the number of covariates in the dimension reduction}

\item{q}{the numbne of study covariates}

\item{I}{a U by U incidence matrix with elements; I(i,j)=1 if state j can be accessed from state i in one step and 0 otherwise}

\item{iniA}{initial value for the iteration}

\item{eps}{the tolerance for convergence, default is 10^-5}
}
\value{
a list of outputs:
\itemize{
\item NewA: the updated A matrix
\item loglikeA: the loglikelihood when updating A
}
}
\description{
This function is used to update A matrix
}
\examples{





}
