% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdata.R
\name{bootdata}
\alias{bootdata}
\title{bootdata}
\usage{
bootdata(data)
}
\arguments{
\item{data}{the original dataset, it has to be a data frame with newid, vdata in format of 1990/9/14, priorstate,currentstate, and covariates}
}
\value{
bootstrap data
}
\description{
This function is used to generate the bootstrap data based on the first status and the empirical probability
It returns a bootstrap data
}
\examples{





}
