% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeB.R
\name{derivativeB}
\alias{derivativeB}
\title{derivativeB}
\usage{
derivativeB(B, ptrans, zy)
}
\arguments{
\item{B}{a numeric coefficient matrix}

\item{ptrans}{a vector with transition number for each state}

\item{zy}{the variable values for a given observation}
}
\value{
loglikelihood
}
\description{
This function is used to calculate the loglikelihood with a given matrix B=AG
}
\examples{





}
