% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\title{derivatives}
\usage{
derivatives(A, Gamma, Dmat, ptrans, zy)
}
\arguments{
\item{A}{matrix with value from previous iteration}

\item{Gamma}{G matrix values}

\item{Dmat}{the coefficient matrix for the fixed variables,}

\item{ptrans}{a vector with transition number for each state}

\item{zy}{the variable values for a given observation}
}
\value{
a list of outputs:
\itemize{
\item fird: the first derivative value
\item secd: the second derivative value
\item loglike: the loglikelihood
}
}
\description{
This function is used calculate the derivative values (first and second derivatives for Newton-Raphson method) and loglikelihood when updating A
}
\examples{





}
