% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(y, ml)
}
\arguments{
\item{y}{the current state}

\item{ml}{the length of the indicator vector}
}
\value{
a indicator vector
}
\description{
This function is used to expand the Y(category) to a indicator vector
}
\examples{





}
