% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gupdate.R
\name{Gupdate}
\alias{Gupdate}
\title{Gupdate}
\usage{
Gupdate(A, Gdata, p, q, I, refG)
}
\arguments{
\item{A}{numeric matrix}

\item{Gdata}{the dataset used to update G}

\item{p}{the number of covariates in the dimension reduction}

\item{q}{the numbne of study covariates}

\item{I}{a U by U incidence matrix with elements; I(i,j)=1 if state j can be accessed from state i in one step and 0 otherwise}

\item{refG}{a vector of reference categories}
}
\value{
a list of outputs:
\itemize{
\item NewG: the updated G matrix
\item loglikeK: the loglikelihood when updating G
\item sderr: standard errors for the coefficient matrix
}
}
\description{
This function is used to update G matrix
}
\examples{





}
