% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\title{norm}
\usage{
norm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a normalized vector with length 1
}
\description{
This function is used to normalize a vector to have unit length
}
