\name{RRMLRfMC}
\alias{RRMLRfMC}
\title{Reduced Rank Multinomial Logistic Regression for Markov Chains}
\usage{
rrmultinom()
}
\description{
Fit the reduced-rank multinomial logistic regression for Markov chain
}
\examples{
rrmultinom(data,rrfactor,ffactor=NULL,intercept=0, priorstate,currentstate, R, Gamma.start, eps = 1e-5)
}
