% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmultinom.R
\name{rrmultinom}
\alias{rrmultinom}
\title{rrmultinom}
\usage{
rrmultinom(
  data,
  rrfactor,
  ffactor = NULL,
  intercept = 0,
  priorstate,
  currentstate,
  R,
  Gamma.start,
  eps = 1e-05
)
}
\arguments{
\item{data}{the dataset}

\item{rrfactor}{the index for variables involved in reduced rank, it could be a numeric vector or a vector with variable name}

\item{ffactor}{the index for fixed variables, it could be a numeric vector or a vector with variable name}

\item{intercept}{used to indicate whether intercept should be considered, 0 means no intercept and 1 means with intercept
\itemize{
\item 0 means to exclude the intercept in the model
\item 1 means to include the intercept in the model
}}

\item{priorstate}{the column index for priorstate}

\item{currentstate}{the column index for currentstate}

\item{R}{the rank}

\item{Gamma.start}{the initial value for G}

\item{eps}{the tolerance for convergence; the default is 10^-5}
}
\value{
a list of outputs:
\itemize{
\item Alpha: the final A matrix
\item Gamma: the final G matrix
\item Beta: the coefficient matrix for variables involved in reduced rank
\item Dcoe: the coefficient matrix for the fixed variables
\item Dsderr: the standard error matrix for the fixed variables
\item Dpval: the p-value matrix for the fixed variables
\item coemat: the overall coefficient matrix
\item niter: the iteration number to get converged
\item df: the degrees of freedom
\item loglik: the final loglikelihood
}
}
\description{
This function is used to fit the reduced rank multinomial logistic regression for markov chain
}
\examples{





}
