% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfun.R
\name{sdfun}
\alias{sdfun}
\title{sdfun}
\usage{
sdfun(
  dat,
  M,
  rank,
  rrfactor = c(3, 4, 6:11),
  ffactor = c(12, 5),
  priorstate = 1,
  currentstate = 2,
  intercept = 0
)
}
\arguments{
\item{dat}{the dataset}

\item{M}{the bootstrap number}

\item{rank}{the rank of the reduced rank model}

\item{rrfactor}{the index for variables in reduced rank part, it can be a numeric vector or a vector with the variable name}

\item{ffactor}{the index for variables in fixed part, it can be a numeric vector or a vector with the variable name}

\item{priorstate}{the column index number for the priorstate, it is a numeric number}

\item{currentstate}{the column index number for the priorstate, it is a numeric number}

\item{intercept}{this parameter indicates whether the intercept is included in the model, it has options 0 or 1 with default as 0
\itemize{
\item 0 means to exclude the intercept in the model
\item 1 means to include the intercept in the model
}}
}
\value{
a list contains resultand matrices with \strong{sd} as the standard error matrix and \strong{pvalue} as the p-value matrix
}
\description{
This function is used get the standard error matrix from bootstrap method
It returns the matrices of standard error and p-value for the coefficient matrix
}
\examples{





}
\references{

}
