% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aupdate.R
\name{Aupdate}
\alias{Aupdate}
\title{Aupdate}
\usage{
Aupdate(Dfix, Gamma, Adata, iniA, R, ptrans, eps)
}
\arguments{
\item{Dfix}{the coefficient matrix for fixed variables}

\item{Gamma}{the G matrix value}

\item{Adata}{the dataset}

\item{iniA}{the initial A matrix; if missing, a constant matrix with elements 0.01}

\item{R}{the rank of reduced rank model}

\item{ptrans}{a vector with transition number for each state}

\item{eps}{the tolerance for convergence, default is 10^-5}
}
\value{
a list of outputs:
\itemize{
\item NewA: the updated A matrix
\item loglikeA: the loglikelihood when updating A
}
}
\description{
This function is used to update A matrix
}
\examples{





}
