% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeB.R
\name{derivativeB}
\alias{derivativeB}
\title{derivativeB}
\usage{
derivativeB(B, I, zy)
}
\arguments{
\item{B}{a numeric coefficient matrix}

\item{I}{U by U incidence matrix with elements; I(i,j)=1 if state j can be accessed from state i in one step and 0 otherwise}

\item{zy}{the variable values for a given observation}
}
\value{
loglikelihood
}
\description{
This function is used to calculate the loglikelihood with a given matrix B=AG
}
\examples{





}
