% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\title{derivatives}
\usage{
derivatives(A, Gamma, Dmat, I, zy, refA)
}
\arguments{
\item{A}{matrix with value from previous iteration}

\item{Gamma}{G matrix values}

\item{Dmat}{the coefficient matrix for the fixed variables,}

\item{I}{a U by U incidence matrix with elements; I(i,j)=1 if state j can be
accessed from state i in one step and 0 otherwise}

\item{zy}{the variable values for a given observation}

\item{refA}{a vector of reference categories}
}
\value{
a list of outputs:
\itemize{
\item fird: the first derivative value
\item secd: the second derivative value
\item loglike: the loglikelihood
}
}
\description{
This function is used calculate the derivative values (first and second derivatives
    for Newton-Raphson method) and loglikelihood when updating A
}
