% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(pri, curr, I, refE)
}
\arguments{
\item{pri}{the prior state}

\item{curr}{the current state}

\item{I}{a U by U incidence matrix with elements; I(i,j)=1 if state j can be accessed from state i in one step and 0 otherwise}

\item{refE}{a vector with the reference categories}
}
\value{
ry: a indicator vector
}
\description{
This function is used to expand the Y(category) to a indicator vector
}
\examples{





}
