% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gupdate.R
\name{Gupdate}
\alias{Gupdate}
\title{Gupdate}
\usage{
Gupdate(A, Gdata, ptrans)
}
\arguments{
\item{A}{numeric matrix}

\item{Gdata}{the dataset used to update G}

\item{ptrans}{a vector with transition number for each state}
}
\value{
a list of outputs:
\itemize{
\item NewG: the updated G matrix
\item loglikeK: the loglikelihood when updating G
\item sderr: standard errors for the coefficient matrix
}
}
\description{
This function is used to update G matrix
}
\examples{





}
